ExchangeAPI/requirements.txt
Flask==1.0.2
schedule==0.5.0
requests==2.20.1
sqlalchemy == 1.2.14
pyodbc == 4.0.24
python-dateutil == 2.7.5
